Moved equipment parent classes under their respective folders
