+++ src/items/equipment/axes/steel-axe.js
@@ -1,4 +1,4 @@
-import Axe from "../axe.js";
+import Axe from "./axe.js";
 
 export default class SteelAxe extends Axe {
     // Attack bonuses
