+++ src/items/equipment/bows/oak-shortbow.js
@@ -1,4 +1,4 @@
-import Bow from "../bow.js";
+import Bow from "./bow.js";
 
 export default class OakShortbow extends Bow {
     // Attack Bonuses
