+++ src/items/equipment/bows/steel-crossbow.js
@@ -1,4 +1,4 @@
-import Bow from "../bow.js";
+import Bow from "./bow.js";
 
 export default class SteelCrossbow extends Bow {
     // Attack Bonuses
