+++ src/items/equipment/daggers/bronze-dagger.js
@@ -1,4 +1,4 @@
-import Dagger from "../dagger.js";
+import Dagger from "./dagger.js";
 
 export default class BronzeDagger extends Dagger {
     // Bonuses
