+++ src/items/equipment/scimitars/bronze-scimitar.js
@@ -1,4 +1,4 @@
-import Scimitar from "../scimitar.js";
+import Scimitar from "./scimitar.js";
 
 export default class BronzeScimitar extends Scimitar {
     // Attack bonuses
