+++ src/items/equipment/staves/air-staff.js
@@ -1,4 +1,4 @@
-import Staff from "../staff.js";
+import Staff from "./staff.js";
 
 export default class AirStaff extends Staff {
     // Attack bonuses
