+++ src/items/equipment/swords/steel-sword.js
@@ -1,4 +1,4 @@
-import Sword from "../sword.js";
+import Sword from "./sword.js";
 
 export default class SteelSword extends Sword {
     // Attack bonuses
